feat: add new judge result 'OLE' (#41)
* add OLE

* some need tobe change

* change output

* all finished

* update detail

* return max value between cpuTime and realTime when TLE

* code style

* update code style settings

* polish

+++ sduoj-judger-interface/src/main/java/cn/edu/sdu/qd/oj/sandbox/enums/SandboxArgument.java
@@ -19,7 +19,7 @@ public enum SandboxArgument {
     MAX_MEMORY("max_memory", Long.class),
     MAX_STACK("max_stack", Long.class),
     MAX_PROCESS_NUMBER("max_process_number", Integer.class),
-    MAX_OUTPUT_SIZE("max_output_size", Integer.class),
+    MAX_OUTPUT_SIZE("max_output_size", Long.class),
     UID("uid", Integer.class),
     GID("gid", Integer.class),
